+++ controller/signal.go
@@ -1,60 +1,68 @@
+/*
+Copyright 2018 BlackRock, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+	http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
 package controller
 
 import (
-	"errors"
+	"context"
+	"fmt"
+	"io"
+	"time"
 
+	"github.com/argoproj/argo-events/common"
 	"github.com/argoproj/argo-events/pkg/apis/sensor/v1alpha1"
-	"github.com/argoproj/argo-events/shared"
-	"github.com/argoproj/argo-events/signals/artifact"
-	"github.com/argoproj/argo-events/signals/calendar"
-	"github.com/argoproj/argo-events/signals/resource"
-	"github.com/argoproj/argo-events/signals/webhook"
+	"github.com/argoproj/argo-events/sdk"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/wait"
 )
-
-var (
-	ErrInvalidSignalType = errors.New("signal type unknown")
-)
-
-// signalCtx is the context for handling signal
-type signalCtx struct {
-	nodename string
-	obj      *v1alpha1.Signal
-	plugin   shared.Signaler
-}
 
 func (soc *sOperationCtx) processSignal(signal v1alpha1.Signal) (*v1alpha1.NodeStatus, error) {
 	soc.log.Debugf("evaluating signal '%s'", signal.Name)
 	node := soc.getNodeByName(signal.Name)
-	if node != nil && node.IsComplete() {
-		soc.stopSignal(&signal)
-		return node, nil
+	if node != nil && node.Phase == v1alpha1.NodePhaseComplete {
+		return node, soc.controller.stopSignal(node.ID)
 	}
 
 	if node == nil {
 		node = soc.initializeNode(signal.Name, v1alpha1.NodeTypeSignal, v1alpha1.NodePhaseNew)
 	}
 
-	// if signal is new (phase=new) == signal not present (signaler=nil)
 	if node.Phase == v1alpha1.NodePhaseNew {
-		// can the node phase and the status of the signal watch ever introduce discrepency?
-		if !soc.signalIsPresent(signal.Name) {
+		if !soc.signalIsPresent(node.ID) {
+			// under normal operations, the signal stream is not present when the node is new
+			// we now attempt to establish a bi-directional stream with our signal microservice
+			// and watch the stream for events.
 			err := soc.watchSignal(&signal)
 			if err != nil {
 				return nil, err
 			}
 		} else {
-			// we get here if the phase == New and the signaler already exists -- should never happen...
-			// in any case, let's log a warning as this is unintended but let's keep the signaler running to receive events
-			// TODO: add check to see if plugin is running?
-			soc.log.Warn("signal is new however signal plugin is already present")
+			// we get here if the stream already exists even though the node phase is new.
+			// this can happen if the sensor was deleted and re-created.
+			// let's log a warning but let's keep the stream running to receive events.
+			// TODO: add check to see if service is running and we're reading from the stream?
+			soc.log.Info("WARNING: signal '%s' is new however signal stream is already present", signal.Name)
 		}
 	} else {
-		if !soc.signalIsPresent(signal.Name) {
-			// signal is active but signaler plugin is missing... meaning we could have been missing events for this signal
-			// we could get here if the controller crashes...
-			// plugin may have crashed?
-			soc.log.Warn("signal is active however signal plugin is missing, potentially could have missed events! attempting to restart plugin.")
+		if !soc.signalIsPresent(node.ID) {
+			// we get here if the signal is in active or error phase and the signal stream is missing.
+			// this means that we had down-time for this signal - we could have missed events.
+			// this can happen if the controller or a signal pod serving the stream goes down.
+			// let's log a warning and attempt to re-establish a stream and watch for events.
+			soc.log.Infof("WARNING: event stream for signal '%s' is missing - could have missed events! reconnecting stream...", signal.Name)
 			err := soc.watchSignal(&signal)
 			if err != nil {
 				return nil, err
@@ -72,125 +80,163 @@ func (soc *sOperationCtx) processSignal(signal v1alpha1.Signal) (*v1alpha1.NodeS
 		return soc.markNodePhase(signal.Name, v1alpha1.NodePhaseComplete), nil
 	}
 
-	return soc.markNodePhase(signal.Name, v1alpha1.NodePhaseActive), nil
+	return soc.markNodePhase(signal.Name, v1alpha1.NodePhaseActive, "stream established"), nil
 }
 
 // checks to see if the signal is present
-// TODO: include a check on the signaler interface to check if signal is active?
-func (soc *sOperationCtx) signalIsPresent(name string) bool {
-	nodeID := soc.s.NodeID(name)
+// TODO: include a check on the stream interface to check if stream is still open?
+func (soc *sOperationCtx) signalIsPresent(nodeID string) bool {
 	soc.controller.signalMu.Lock()
-	_, ok := soc.controller.signals[nodeID]
+	_, ok := soc.controller.signalStreams[nodeID]
 	soc.controller.signalMu.Unlock()
 	return ok
 }
 
-// resolveSignaler is a helper method to find the correct Signaler for this signal definition
-func (soc *sOperationCtx) resolveSignaler(signal *v1alpha1.Signal) (shared.Signaler, error) {
-	var signaler shared.Signaler
+// resolveClient is a helper method to find the correct SignalClient for this signal
+func (soc *sOperationCtx) resolveClient(signal *v1alpha1.Signal) (sdk.SignalClient, error) {
+	var typ string
 	switch signal.GetType() {
 	case v1alpha1.SignalTypeStream:
-		// these come from plugins
-		plugin, err := soc.controller.pluginMgr.Dispense(signal.Stream.Type)
-		if err != nil {
-			return nil, err
-		}
-		signaler = plugin.(shared.Signaler)
-	case v1alpha1.SignalTypeArtifact:
-		streamPlugin, err := soc.controller.pluginMgr.Dispense(signal.Artifact.Target.Type)
-		if err != nil {
-			return nil, err
-		}
-		streamSignaler := streamPlugin.(shared.Signaler)
-		signaler = artifact.New(streamSignaler, soc.controller.kubeClientset, soc.controller.Config.Namespace)
-	case v1alpha1.SignalTypeCalendar:
-		signaler = calendar.New()
-	case v1alpha1.SignalTypeResource:
-		signaler = resource.New(soc.controller.kubeConfig)
-	case v1alpha1.SignalTypeWebhook:
-		signaler = webhook.New()
+		typ = signal.Stream.Type
 	default:
-		return nil, ErrInvalidSignalType
+		typ = string(signal.GetType())
 	}
-	return signaler, nil
+	client, err := soc.controller.signalMgr.Dispense(typ)
+	if err != nil {
+		return nil, err
+	}
+	return client, nil
 }
 
 // adds a new signal to the controller's signals
 func (soc *sOperationCtx) watchSignal(signal *v1alpha1.Signal) error {
-	signaler, err := soc.resolveSignaler(signal)
+	client, err := soc.resolveClient(signal)
 	if err != nil {
 		return err
 	}
 
-	events, err := signaler.Start(signal)
+	// create the context for this stream
+	ctx := context.Background()
+	var cancel context.CancelFunc
+	if signal.Deadline > 0 {
+		ctx, cancel = context.WithTimeout(ctx, time.Duration(signal.Deadline))
+	} else {
+		ctx, cancel = context.WithCancel(ctx)
+	}
+
+	stream, err := client.Listen(ctx, signal)
 	if err != nil {
+		cancel()
 		return err
 	}
 	nodeID := soc.s.NodeID(signal.Name)
 	soc.controller.signalMu.Lock()
-	soc.controller.signals[nodeID] = signaler
+	soc.controller.signalStreams[nodeID] = stream
 	soc.controller.signalMu.Unlock()
 
-	go soc.controller.listenForEvents(soc.s.Name, signal.Name, nodeID, events)
+	go soc.controller.listenOnStream(soc.s.Name, signal.Name, nodeID, stream, cancel)
 	return nil
 }
 
-func (soc *sOperationCtx) stopSignal(signal *v1alpha1.Signal) {
-	nodeID := soc.s.NodeID(signal.Name)
-	soc.controller.signalMu.Lock()
-	signaler, ok := soc.controller.signals[nodeID]
-	soc.controller.signals[nodeID] = nil
-	delete(soc.controller.signals, nodeID)
-	soc.controller.signalMu.Unlock()
-	if ok && signaler != nil {
-		if err := signaler.Stop(); err != nil {
-			soc.log.Warnf("failed to stop the signaler '%s'. cause: %s", nodeID, err)
+// stop the signal by:
+// 1. deleting the stream from the controller's signalStreams map
+// 2. sending the terminate signal on the stream and close it
+// NOTE: this is a method on the controller
+func (c *SensorController) stopSignal(nodeID string) error {
+	c.signalMu.Lock()
+	stream, ok := c.signalStreams[nodeID]
+	c.signalStreams[nodeID] = nil
+	delete(c.signalStreams, nodeID)
+	c.signalMu.Unlock()
+	if ok && stream != nil {
+		if err := stream.Send(sdk.Terminate); err != nil {
+			return err
+		}
+		if err := stream.Close(); err != nil {
+			return err
 		}
 	}
+	return nil
 }
 
-// listens for events on the events channel
-// meant to be run as a separate goroutine
-// this will terminate once the events chan is closed
-// the events chan should be closed by the future operations on this sensor
-func (c *SensorController) listenForEvents(sensor, signal, nodeID string, events <-chan *v1alpha1.Event) {
+// listens for events on the event stream. meant to be run as a separate goroutine
+// this will terminate once the stream receives an EOF indicating it has completed or it encounters a stream error.
+// NOTE: this is a method on the controller
+func (c *SensorController) listenOnStream(sensor, signal, nodeID string, stream sdk.SignalService_ListenService, cancel context.CancelFunc) {
+	// TODO: possible context leak if we don't utilize cancel
+	//defer cancel()
 	sensors := c.sensorClientset.ArgoprojV1alpha1().Sensors(c.Config.Namespace)
-	for event := range events {
-		c.log.Infof("sensor '%s' received event for signal '%s'", sensor, signal)
-		// todo: store the event in an external store
+	for {
+		in, streamErr := stream.Recv()
+		if streamErr == io.EOF {
+			return
+		}
 
+		// retrieve the sensor & node as this will be needed to update accordingly
 		s, err := sensors.Get(sensor, metav1.GetOptions{})
 		if err != nil {
-			// we can get here if the sensor was removed after completion or deleted and the signaler was not stopped
-			// we should attempt to log a warning and stop the signaler
-			c.log.Warnf("sensor '%s' cannot be found due to: %s. Stopping signaler...", sensor, err)
-			c.signalMu.Lock()
-			defer c.signalMu.Unlock()
-			signaler, ok := c.signals[nodeID]
-			if !ok {
-				c.log.Panicf("signaler '%s' not found, unable to stop listenForEvents() gracefully", nodeID)
-			}
-			err := signaler.Stop()
+			// we can get here if the sensor was deleted and the stream was not closed
+			// we should attempt to log a warning and stop & close the signal stream
+			// TODO: why call stopSignal() when we have access to the stream within this func?
+			c.log.Warnf("sensor '%s' cannot be found due to: %s. Stopping signal '%s' event stream...", sensor, err, signal)
+			err := c.stopSignal(nodeID)
 			if err != nil {
-				c.log.Panicf("failed to stop signaler '%s'", nodeID)
+				c.log.Panicf("failed to stop signal stream '%s': %s", nodeID, err)
 			}
 		}
-		// requeue this sensor with the event context
-		/*
-			key, err := cache.MetaNamespaceKeyFunc(s)
-			if err == nil {
-				c.queue.Add(key)
-			}
-		*/
+		phase := s.Status.Phase
+		msg := s.Status.Message
 		node, ok := s.Status.Nodes[nodeID]
 		if !ok {
+			// TODO: should we re-initialize this node?
 			c.log.Panicf("'%s' node is missing from sensor's nodes", nodeID)
 		}
-		node.LatestEvent = &v1alpha1.EventWrapper{Event: *event}
-		s.Status.Nodes[nodeID] = node
-		_, err = sensors.Update(s)
+
+		if streamErr != nil {
+			c.log.Infof("ERROR: sensor '%s' signal '%s' stream failed: %s", sensor, signal, streamErr)
+			// error received from the stream
+			// remove the stream from the signalStreams map
+			c.signalMu.Lock()
+			c.signalStreams[nodeID] = nil
+			delete(c.signalStreams, nodeID)
+			c.signalMu.Unlock()
+
+			// mark the sensor & node as error phase
+			phase = v1alpha1.NodePhaseError
+			msg = fmt.Sprintf("signal '%s' encountered stream err: %s", signal, streamErr)
+			node.Phase = v1alpha1.NodePhaseError
+			node.Message = streamErr.Error()
+		} else {
+			c.log.Debugf("sensor '%s' received event for signal '%s'", sensor, signal)
+			node.LatestEvent = &v1alpha1.EventWrapper{Event: *in.Event}
+		}
+
+		// TODO: perform a Patch here instead as the sensor could become stale
+		// and this exponential backoff is slow
+		err = wait.ExponentialBackoff(common.DefaultRetry, func() (bool, error) {
+			s, err := sensors.Get(sensor, metav1.GetOptions{})
+			if err != nil {
+				return false, err
+			}
+			s.Status.Nodes[nodeID] = node
+			s.Status.Phase = phase
+			s.Status.Message = msg
+			_, err = sensors.Update(s)
+			if err != nil {
+				if !common.IsRetryableKubeAPIError(err) {
+					return false, err
+				}
+				return false, nil
+			}
+			return true, nil
+		})
 		if err != nil {
-			c.log.Panicf("failed to update sensor with event")
+			c.log.Panicf("failed to update sensor: %s", err)
+		}
+
+		// finally check if there was a streamErr, we must return
+		if streamErr != nil {
+			return
 		}
 	}
 }