+++ shared/doc.go
@@ -1,0 +1,4 @@
+// Package shared contains the necessary files for registering
+// a micro service for the sensor controller
+// This package should not be built as part of a signal server!
+package shared