+++ signals/stream/builtin/amqp/amqp.go
@@ -14,14 +14,14 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package amqp
 
 import (
 	"fmt"
+	"log"
 
 	"github.com/argoproj/argo-events/pkg/apis/sensor/v1alpha1"
-	"github.com/argoproj/argo-events/shared"
-	plugin "github.com/hashicorp/go-plugin"
+	"github.com/argoproj/argo-events/sdk"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	amqplib "github.com/streadway/amqp"
@@ -35,98 +35,107 @@ const (
 	routingKeyKey   = "routingKey"
 )
 
-type amqp struct {
-	conn     *amqplib.Connection
-	delivery <-chan amqplib.Delivery
+// Note: micro requires stateless operation so the Listen() method should not use the
+// receive struct to save or modify state.
+type amqp struct{}
+
+// New creates an amqp listener
+func New() sdk.Listener {
+	return new(amqp)
 }
 
-// New creates an amqp signaler
-func New() shared.Signaler {
-	return &amqp{}
+func (*amqp) Listen(signal *v1alpha1.Signal, done <-chan struct{}) (<-chan *v1alpha1.Event, error) {
+	conn, err := amqplib.Dial(signal.Stream.URL)
+	if err != nil {
+		return nil, fmt.Errorf("failed to connect to server: %s", err)
+	}
+	ch, err := conn.Channel()
+	if err != nil {
+		return nil, fmt.Errorf("failed to open channel: %s", err)
+	}
+
+	delivery, err := getDelivery(ch, signal.Stream.Attributes)
+	events := make(chan *v1alpha1.Event)
+
+	// start listening for messages
+	go func() {
+		defer close(events)
+		for {
+			select {
+			case msg := <-delivery:
+				event := &v1alpha1.Event{
+					Context: v1alpha1.EventContext{
+						EventID:            msg.MessageId,
+						EventType:          EventType,
+						EventTypeVersion:   msg.ConsumerTag,
+						CloudEventsVersion: sdk.CloudEventsVersion,
+						Source:             &v1alpha1.URI{},
+						EventTime:          metav1.Time{Time: msg.Timestamp},
+						SchemaURL:          &v1alpha1.URI{},
+						ContentType:        msg.ContentType,
+						Extensions:         map[string]string{"content-encoding": msg.ContentEncoding},
+					},
+					Data: msg.Body,
+				}
+				events <- event
+				msg.Ack(false)
+			case <-done:
+				if err := ch.Close(); err != nil {
+					log.Printf("failed to close channel for signal '%s': %s", signal.Name, err)
+				}
+				if err := conn.Close(); err != nil {
+					log.Panicf("failed to close connection for signal '%s': %s", signal.Name, err)
+				}
+				log.Printf("shut down signal '%s'", signal.Name)
+				return
+			}
+		}
+	}()
+
+	return events, nil
 }
 
-func (a *amqp) Start(signal *v1alpha1.Signal) (<-chan *v1alpha1.Event, error) {
-	// parse out the attributes
-	exchangeName, ok := signal.Stream.Attributes[exchangeNameKey]
-	if !ok {
-		return nil, shared.ErrMissingRequiredAttribute
-	}
-	exchangeType, ok := signal.Stream.Attributes[exchangeTypeKey]
-	if !ok {
-		return nil, shared.ErrMissingRequiredAttribute
-	}
-	routingKey, ok := signal.Stream.Attributes[routingKeyKey]
-	if !ok {
-		return nil, shared.ErrMissingRequiredAttribute
+func getDelivery(ch *amqplib.Channel, attr map[string]string) (<-chan amqplib.Delivery, error) {
+	exName, exType, rKey, err := parseAttributes(attr)
+	if err != nil {
+		return nil, err
 	}
 
-	var err error
-	a.conn, err = amqplib.Dial(signal.Stream.URL)
+	err = ch.ExchangeDeclare(exName, exType, true, false, false, false, nil)
 	if err != nil {
-		return nil, fmt.Errorf("failed to connect to RabbitMQ. cause: %s", err)
-	}
-
-	ch, err := a.conn.Channel()
-	if err != nil {
-		return nil, fmt.Errorf("failed to open channel. cause: %s", err)
-	}
-
-	err = ch.ExchangeDeclare(exchangeName, exchangeType, true, false, false, false, nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed to declare RabbitMQ %s exchange '%s'. cause: %s", exchangeType, exchangeName, err)
+		return nil, fmt.Errorf("failed to declare %s exchange '%s': %s", exType, exName, err)
 	}
 
 	q, err := ch.QueueDeclare("", false, false, true, false, nil)
 	if err != nil {
-		return nil, fmt.Errorf("failed to declare RabbigMQ queue. cause: %s", err)
+		return nil, fmt.Errorf("failed to declare queue: %s", err)
 	}
 
-	err = ch.QueueBind(q.Name, routingKey, exchangeName, false, nil)
+	err = ch.QueueBind(q.Name, rKey, exName, false, nil)
 	if err != nil {
-		return nil, fmt.Errorf("failed to bind RabbitMQ %s exchange '%s' to queue with routingKey: %s. cause: %s", exchangeType, exchangeName, routingKey, err)
+		return nil, fmt.Errorf("failed to bind %s exchange '%s' to queue with routingKey: %s: %s", exType, exName, rKey, err)
 	}
-	a.delivery, err = ch.Consume(q.Name, "", true, false, false, false, nil)
+
+	delivery, err := ch.Consume(q.Name, "", true, false, false, false, nil)
 	if err != nil {
-		return nil, fmt.Errorf("failed to begin consuming RabbitMQ messages. cause: %s", err)
+		return nil, fmt.Errorf("failed to begin consuming messages: %s", err)
 	}
-
-	events := make(chan *v1alpha1.Event)
-	go a.listen(events)
-	return events, nil
+	return delivery, nil
 }
 
-func (a *amqp) Stop() error {
-	return a.conn.Close()
-}
-
-func (a *amqp) listen(events chan *v1alpha1.Event) {
-	for msg := range a.delivery {
-		event := &v1alpha1.Event{
-			Context: v1alpha1.EventContext{
-				EventID:            msg.MessageId,
-				EventType:          EventType,
-				EventTypeVersion:   msg.ConsumerTag,
-				CloudEventsVersion: shared.CloudEventsVersion,
-				Source:             &v1alpha1.URI{},
-				EventTime:          metav1.Time{Time: msg.Timestamp},
-				SchemaURL:          &v1alpha1.URI{},
-				ContentType:        msg.ContentType,
-				Extensions:         map[string]string{"content-encoding": msg.ContentEncoding},
-			},
-			Data: msg.Body,
-		}
-		events <- event
-		msg.Ack(false)
+func parseAttributes(attr map[string]string) (string, string, string, error) {
+	// parse out the attributes
+	exchangeName, ok := attr[exchangeNameKey]
+	if !ok {
+		return "", "", "", sdk.ErrMissingRequiredAttribute
 	}
-}
-
-func main() {
-	amqp := New()
-	plugin.Serve(&plugin.ServeConfig{
-		HandshakeConfig: shared.Handshake,
-		Plugins: map[string]plugin.Plugin{
-			shared.SignalPluginName: shared.NewPlugin(amqp),
-		},
-		GRPCServer: plugin.DefaultGRPCServer,
-	})
+	exchangeType, ok := attr[exchangeTypeKey]
+	if !ok {
+		return exchangeName, "", "", sdk.ErrMissingRequiredAttribute
+	}
+	routingKey, ok := attr[routingKeyKey]
+	if !ok {
+		return exchangeName, "", "", sdk.ErrMissingRequiredAttribute
+	}
+	return exchangeName, exchangeType, routingKey, nil
 }