+++ signals/stream/builtin/kafka/kafka.go
@@ -14,16 +14,16 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package kafka
 
 import (
 	"fmt"
+	"log"
 	"strconv"
 
 	"github.com/Shopify/sarama"
 	"github.com/argoproj/argo-events/pkg/apis/sensor/v1alpha1"
-	"github.com/argoproj/argo-events/shared"
-	plugin "github.com/hashicorp/go-plugin"
+	"github.com/argoproj/argo-events/sdk"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
@@ -33,119 +33,109 @@ const (
 	EventType    = "org.apache.kafka.pub"
 )
 
-type kafka struct {
-	consumer          sarama.Consumer
-	partitionConsumer sarama.PartitionConsumer
-	stop              chan struct{}
-}
+// Note: micro requires stateless operation so the Listen() method should not use the
+// receive struct to save or modify state.
+type kafka struct{}
 
 // New creates a new kafka signaler
-func New() shared.Signaler {
-	return &kafka{
-		stop: make(chan struct{}),
-	}
+func New() sdk.Listener {
+	return new(kafka)
 }
 
-func (k *kafka) Start(signal *v1alpha1.Signal) (<-chan *v1alpha1.Event, error) {
+func (*kafka) Listen(signal *v1alpha1.Signal, done <-chan struct{}) (<-chan *v1alpha1.Event, error) {
 	// parse out the attributes
-	topic, ok := signal.Stream.Attributes["topic"]
-	if !ok {
-		return nil, shared.ErrMissingRequiredAttribute
-	}
-	var pString string
-	if pString, ok = signal.Stream.Attributes[partitionKey]; !ok {
-		return nil, fmt.Errorf(shared.ErrMissingAttribute, partitionKey)
-	}
-	pInt, err := strconv.ParseInt(pString, 10, 32)
+	topic, partition, err := parseAttributes(signal.Stream.Attributes)
 	if err != nil {
 		return nil, err
 	}
-	partition := int32(pInt)
 
-	k.consumer, err = sarama.NewConsumer([]string{signal.Stream.URL}, nil)
+	consumer, err := sarama.NewConsumer([]string{signal.Stream.URL}, nil)
 	if err != nil {
-		return nil, fmt.Errorf("failed to connect to kafka cluster at %s", signal.Stream.URL)
+		return nil, fmt.Errorf("failed to connect to cluster at %s", signal.Stream.URL)
 	}
 
-	availablePartitions, err := k.consumer.Partitions(topic)
+	availablePartitions, err := consumer.Partitions(topic)
 	if err != nil {
 		return nil, fmt.Errorf("unable to get available partitions for kafka topic '%s'. cause: %s", topic, err.Error())
 	}
-
-	if ok := k.verifyPartitionAvailable(partition, availablePartitions); !ok {
+	if ok := verifyPartitionAvailable(partition, availablePartitions); !ok {
 		return nil, fmt.Errorf("partition %v does not exist for topic '%s'", partition, topic)
 	}
 
-	k.partitionConsumer, err = k.consumer.ConsumePartition(topic, partition, sarama.OffsetNewest)
+	partitionConsumer, err := consumer.ConsumePartition(topic, partition, sarama.OffsetNewest)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create partition consumer for topic '%s' and partition: %v. cause: %s", topic, partition, err.Error())
 	}
+
 	events := make(chan *v1alpha1.Event)
-	go k.listen(events)
+
+	// start listening for messages
+	go func() {
+		defer close(events)
+		for {
+			select {
+			case msg := <-partitionConsumer.Messages():
+				event := &v1alpha1.Event{
+					Context: v1alpha1.EventContext{
+						EventID:            fmt.Sprintf("partition-%v-offset-%v", msg.Partition, msg.Offset),
+						EventType:          EventType,
+						EventTime:          metav1.Time{Time: msg.Timestamp},
+						CloudEventsVersion: sdk.CloudEventsVersion,
+						Extensions:         make(map[string]string),
+					},
+					Data: msg.Value,
+				}
+				log.Printf("signal '%s' received msg", signal.Name)
+				events <- event
+			case err := <-partitionConsumer.Errors():
+				event := &v1alpha1.Event{
+					Context: v1alpha1.EventContext{
+						EventID:            fmt.Sprintf("partition-%v-", err.Partition),
+						EventType:          EventType,
+						CloudEventsVersion: sdk.CloudEventsVersion,
+						Extensions:         map[string]string{sdk.ContextExtensionErrorKey: err.Err.Error()},
+					},
+				}
+				log.Printf("signal '%s' received error", signal.Name)
+				events <- event
+			case <-done:
+				if err := partitionConsumer.Close(); err != nil {
+					log.Printf("failed to close partition consumer for signal '%s': %s", signal.Name, err)
+				}
+				if err := consumer.Close(); err != nil {
+					log.Panicf("failed to close consumer for signal '%s': %s", signal.Name, err)
+				}
+				log.Printf("shut down signal '%s'", signal.Name)
+				return
+			}
+		}
+	}()
+
 	return events, nil
 }
 
-func (k *kafka) Stop() error {
-	k.stop <- struct{}{}
-	err := k.partitionConsumer.Close()
-	if err != nil {
-		return err
+func parseAttributes(attr map[string]string) (string, int32, error) {
+	topic, ok := attr["topic"]
+	if !ok {
+		return "", 0, sdk.ErrMissingRequiredAttribute
 	}
-	err = k.consumer.Close()
-	if err != nil {
-		return err
+	var pString string
+	if pString, ok = attr[partitionKey]; !ok {
+		return topic, 0, fmt.Errorf(sdk.ErrMissingAttribute, partitionKey)
 	}
-	return nil
+	pInt, err := strconv.ParseInt(pString, 10, 32)
+	if err != nil {
+		return topic, 0, err
+	}
+	partition := int32(pInt)
+	return topic, partition, nil
 }
 
-func (k *kafka) listen(events chan *v1alpha1.Event) {
-	defer close(events)
-	for {
-		select {
-		case msg := <-k.partitionConsumer.Messages():
-			event := &v1alpha1.Event{
-				Context: v1alpha1.EventContext{
-					EventID:            fmt.Sprintf("partition-%v-offset-%v", msg.Partition, msg.Offset),
-					EventType:          EventType,
-					EventTime:          metav1.Time{Time: msg.Timestamp},
-					CloudEventsVersion: shared.CloudEventsVersion,
-					Extensions:         make(map[string]string),
-				},
-				Data: msg.Value,
-			}
-			events <- event
-		case err := <-k.partitionConsumer.Errors():
-			event := &v1alpha1.Event{
-				Context: v1alpha1.EventContext{
-					EventID:            fmt.Sprintf("partition-%v-", err.Partition),
-					EventType:          EventType,
-					CloudEventsVersion: shared.CloudEventsVersion,
-					Extensions:         map[string]string{shared.ContextExtensionErrorKey: err.Err.Error()},
-				},
-			}
-			events <- event
-		case <-k.stop:
-			return
-		}
-	}
-}
-
-func (k *kafka) verifyPartitionAvailable(part int32, partitions []int32) bool {
+func verifyPartitionAvailable(part int32, partitions []int32) bool {
 	for _, p := range partitions {
 		if part == p {
 			return true
 		}
 	}
 	return false
-}
-
-func main() {
-	kafka := New()
-	plugin.Serve(&plugin.ServeConfig{
-		HandshakeConfig: shared.Handshake,
-		Plugins: map[string]plugin.Plugin{
-			shared.SignalPluginName: shared.NewPlugin(kafka),
-		},
-		GRPCServer: plugin.DefaultGRPCServer,
-	})
 }