+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
 
-## 1.0.0 (tbd)
+## v0.5-beta1 (tbd)
++ Signals as separate deployments [#49](https://github.com/argoproj/argo-events/pull/49)
++ Fixed code-gen bug [#46](https://github.com/argoproj/argo-events/issues/46)
++ Filters for signals [#26](https://github.com/argoproj/argo-events/issues/26)
+
+## v0.5-alpha1
 + Initial release